@@ -1,6 +1,6 @@
 {
   "name": "ngx.tools",
-  "version": "9.0.0",
+  "version": "13.0.0",
   "license": "MIT",
   "author": {
     "name": "Wassim Chegham",
@@ -11,6 +11,7 @@
     "ng": "ng",
     "start": "ng serve",
     "build": "ng build",
+    "watch": "ng build --watch --configuration development",
     "postbuild": "bash scripts/stamp-build.sh",
     "test": "ng test",
     "lint": "ng lint",
@@ -24,50 +25,43 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/cdk": "^9.0.0",
-    "@angular/http": "^7.2.16",
-    "@angular/material": "^9.0.0",
-    "@angular/pwa": "^0.900.2",
-    "@angular/service-worker": "^9.0.1",
-    "algoliasearch": "^4.0.3",
-    "algoliasearch-helper": "^3.1.0",
-    "core-js": "^3.6.4",
+    "@angular/animations": "~13.3.0",
+    "@angular/cdk": "^13.3.5",
+    "@angular/common": "~13.3.0",
+    "@angular/compiler": "~13.3.0",
+    "@angular/core": "~13.3.0",
+    "@angular/forms": "~13.3.0",
+    "@angular/material": "^13.3.5",
+    "@angular/platform-browser": "~13.3.0",
+    "@angular/platform-browser-dynamic": "~13.3.0",
+    "@angular/pwa": "^13.3.5",
+    "@angular/router": "~13.3.0",
+    "@angular/service-worker": "^13.3.5",
+    "algoliasearch": "^4.13.0",
+    "algoliasearch-helper": "^3.8.2",
     "hammerjs": "^2.0.8",
-    "@angular/animations": "~9.0.1",
-    "@angular/common": "~9.0.1",
-    "@angular/compiler": "~9.0.1",
-    "@angular/core": "~9.0.1",
-    "@angular/forms": "~9.0.1",
-    "@angular/platform-browser": "~9.0.1",
-    "@angular/platform-browser-dynamic": "~9.0.1",
-    "@angular/router": "~9.0.1",
-    "rxjs": "~6.5.4",
-    "tslib": "^1.10.0",
-    "zone.js": "~0.10.2"
+    "rxjs": "~7.5.0",
+    "tslib": "^2.3.0",
+    "zone.js": "~0.11.4"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~0.900.2",
-    "@angular/cli": "~9.0.2",
-    "@angular/compiler-cli": "~9.0.1",
-    "@angular/language-service": "~9.0.1",
-    "@release-it/conventional-changelog": "^1.1.0",
-    "@types/jasmine": "~3.5.0",
-    "@types/jasminewd2": "~2.0.3",
+    "@angular-devkit/build-angular": "~13.3.5",
+    "@angular/cli": "~13.3.5",
+    "@angular/compiler-cli": "~13.3.0",
+    "@angular/language-service": "~13.3.5",
+    "@release-it/conventional-changelog": "^5.0.0",
+    "@types/jasmine": "~3.10.0",
     "@types/node": "^12.11.1",
-    "codelyzer": "^5.1.2",
     "conventional-changelog-cli": "^2.0.31",
-    "jasmine-core": "~3.5.0",
+    "jasmine-core": "~4.0.0",
     "jasmine-spec-reporter": "~4.2.1",
-    "karma": "~4.3.0",
+    "karma": "~6.3.0",
     "karma-chrome-launcher": "~3.1.0",
+    "karma-coverage": "~2.1.0",
     "karma-coverage-istanbul-reporter": "~2.1.0",
-    "karma-jasmine": "~2.0.1",
-    "karma-jasmine-html-reporter": "^1.4.2",
-    "protractor": "~5.4.3",
-    "release-it": "^12.5.0",
-    "terser": "^4.6.3",
-    "ts-node": "~8.3.0",
-    "tslint": "~5.18.0",
-    "typescript": "~3.7.5"
+    "karma-jasmine": "~4.0.0",
+    "karma-jasmine-html-reporter": "~1.7.0",
+    "release-it": "^15.0.0",
+    "typescript": "~4.6.2"
   }
 }
