@@ -4,7 +4,7 @@ import { NotFoundComponent } from './shared/not-found/not-found.component';
 
 export const routes: Routes = [
   { path: '', redirectTo: '/search', pathMatch: 'full' },
-  { path: 'search', loadChildren: './search/search.module#SearchModule' },
+  { path: 'search', loadChildren: () => import('./search/search.module').then(m => m.SearchModule) },
   { path: '**', component: NotFoundComponent }
 ];
 
