@@ -8,12 +8,12 @@ export class MetricsService {
 
   send(results: {query: string}) {
     try {
-      window['ga']('send', {
+      (window as any)['ga']('send', {
         hitType: 'event',
         eventCategory: 'Search',
         eventAction: 'query',
         eventLabel: results.query
       });
-    } catch (e) {}
+    } catch {}
   }
 }
